added/moved the Navbar.jsx changes to NavigationBar.jsx
